Finalização

+++ aula1/Forum/src/main/java/br/com/forum/controller/TopicosController.java
@@ -3,6 +3,7 @@ package br.com.forum.controller;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import javax.transaction.Transactional;
 import javax.validation.Valid;
@@ -18,6 +19,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.util.UriComponentsBuilder;
+
+
 
 import br.com.forum.controller.form.AtualizacaoTopicoForm;
 import br.com.forum.dto.DetalhesDoTopicoDto;
@@ -49,7 +52,6 @@ public class TopicosController {
 				
 	}
 	@PostMapping
-	@Transactional
 	public ResponseEntity<TopicoDto> cadastrar(@RequestBody @Valid TopicoForm form, UriComponentsBuilder uriBuilder) {
 		Topico topico  = form.converter(cursoRepository);
 		topicoRepository.save(topico);
@@ -66,28 +68,37 @@ public class TopicosController {
 	}**/
 	
 	@GetMapping("/{id}") //@PathVariable informa que o variavel vem como parte da url e não como interrogação
-	public DetalhesDoTopicoDto detalhar(@PathVariable Long id) {
-		Topico topico = topicoRepository.getOne(id); // faz a busca por id
-		return new DetalhesDoTopicoDto(topico); // converte em DTo
+	public ResponseEntity<DetalhesDoTopicoDto>  detalhar(@PathVariable Long id) {
+		Optional<Topico> topico = topicoRepository.findById(id);
+		if(topico.isPresent()) {
+			return ResponseEntity.ok(new DetalhesDoTopicoDto(topico.get()));
+		}
+		
+		return ResponseEntity.notFound().build();
 		
 	}
 	
 	@PutMapping("/{id}")
 	@Transactional
 	public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicoForm form  ){
-		Topico topico = form.atualizar(id, topicoRepository);
-		
-		return ResponseEntity.ok(new TopicoDto(topico));
-		
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if(optional.isPresent()) {
+			Topico topico = form.atualizar(id, topicoRepository);
+			return ResponseEntity.ok(new TopicoDto(topico));
+		}
+		return ResponseEntity.notFound().build();
+			
 	}
 	
 	@DeleteMapping("/{id}")
 	@Transactional
 	public  ResponseEntity<?> remover(@PathVariable Long id){
-		topicoRepository.deleteById(id);
-		return ResponseEntity.ok().build();
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if(optional.isPresent()) {
+			topicoRepository.deleteById(id);
+			return ResponseEntity.ok().build();
+		}
+		return ResponseEntity.notFound().build();
 	}
 	
-	
-
 }
